refactor(date): remove dateGenericAdapter in favor of createDateEntity
Removed the dateGenericAdapter as it was only responsible for creating a date entity, which does not reflect the purpose of the Adapter pattern. Replaced it with direct calls to createDateEntity to improve clarity and maintain architectural consistency.
