+++ adapters/cart/prisma/handle-get-prisma-cart.ts
@@ -1,7 +1,6 @@
 import { CartEntity } from "@/domain/entities/cart.entity";
 import { CartDatabase } from "@/infra/db/types/cart";
 import { mapPrismaCartToDomainCart } from "@/adapters/cart/prisma/map-prisma-cart-to-domain-cart";
-import { dateGenericAdapter } from "@/adapters/date/generic/date-generic.adapter";
 
 export function handleGetPrismaCart(
   dbCart: CartDatabase | null
@@ -13,7 +12,7 @@ export function handleGetPrismaCart(
   let cart: CartEntity | null;
 
   try {
-    cart = mapPrismaCartToDomainCart(dbCart, dateGenericAdapter);
+    cart = mapPrismaCartToDomainCart(dbCart);
   } catch (error) {
     console.error("Error while mapping API cart to domain cart: ", error);
     cart = null;
