+++ adapters/product/prisma/get-prisma-latest-products-service.adapter.ts
@@ -1,7 +1,6 @@
 import { ProductEntity } from "@/domain/entities/product.entity";
 import { getPrismaLatestProductsService } from "@/infra/services/product/get-prisma-latest-products.service";
 import { mapPrismaProductToDomainProduct } from "@/adapters/product/prisma/map-prisma-product-to-domain-product";
-import { dateGenericAdapter } from "@/adapters/date/generic/date-generic.adapter";
 
 export async function getPrismaLatestProductsServiceAdapter(): Promise<
   ProductEntity[]
@@ -11,7 +10,7 @@ export async function getPrismaLatestProductsServiceAdapter(): Promise<
 
   try {
     products = response.map((dbProduct) =>
-      mapPrismaProductToDomainProduct(dbProduct, dateGenericAdapter)
+      mapPrismaProductToDomainProduct(dbProduct)
     );
   } catch (error) {
     console.error(
