+++ domain/use-cases/cart/create-new-cart.use-case.ts
@@ -1,12 +1,11 @@
-import { DateAdapter } from "@/adapters/date/date.adapter";
 import { CartEntity } from "@/domain/entities/cart.entity";
 import { createCurrencyEntity } from "@/domain/entities/currency.entity";
+import { createDateEntity } from "@/domain/entities/date.entity";
 import { CartRepository } from "@/domain/repositories/cart.repository";
 import { cartEntitySchema } from "@/lib/validators/cart";
 
 export type FindCartByUserOrSessionCartUseCaseParams = {
   cartRepository: CartRepository;
-  dateAdapter: DateAdapter;
   sessionCartId: string;
   userId?: string;
 };
@@ -14,7 +13,6 @@ export type FindCartByUserOrSessionCartUseCaseParams = {
 export async function createNewCartUseCase({
   cartRepository,
   sessionCartId,
-  dateAdapter,
   userId,
 }: FindCartByUserOrSessionCartUseCaseParams): Promise<CartEntity> {
   const cart: CartEntity = cartEntitySchema.parse({
@@ -26,8 +24,8 @@ export async function createNewCartUseCase({
     shippingPrice: createCurrencyEntity(0),
     taxPrice: createCurrencyEntity(0),
     totalPrice: createCurrencyEntity(0),
-    createdAt: dateAdapter.safeCreateEntity(new Date()),
-    updatedAt: dateAdapter.safeCreateEntity(new Date()),
+    createdAt: createDateEntity(new Date()),
+    updatedAt: createDateEntity(new Date()),
   });
 
   await cartRepository.create(cart);
